more tests; mocked ILogger

+++ CWDocsCore/Services/DocumentService.cs
@@ -1,28 +1,23 @@
-﻿using CWDocsCore.Extensions;
-using CWDocsCore.Models;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
+﻿using CWDocsCore.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using NLog;
-using System.Threading;
 using System.IO;
+using Microsoft.Extensions.Logging;
 
 namespace CWDocsCore.Services {
     public class DocumentService : IDocumentService {
 
         private readonly CWDocsDbContext _context;
-        private readonly Logger _debugLogger;
+        private readonly ILogger _debugLogger;
         private readonly Microsoft.Extensions.Configuration.IConfiguration _settings;
 
 
-        public DocumentService(Microsoft.Extensions.Configuration.IConfiguration settings, CWDocsDbContext context) {
-            _debugLogger = LogManager.GetLogger("debugLogger");
+        public DocumentService(Microsoft.Extensions.Configuration.IConfiguration settings, CWDocsDbContext context, ILogger logger) {
+            //_debugLogger = LogManager.GetLogger("debugLogger");
             _settings = settings;
             _context = context;
+            _debugLogger = logger;
         }
 
         public List<DocumentModel> GetDocuments(UserModel user)
@@ -33,7 +28,7 @@ namespace CWDocsCore.Services {
 
         public DocumentModel GetDocument(UserModel user, int id)
         {
-            DocumentModel document = _context.Documents.Where(d => d.userId == user.Id && d.fileId == id).FirstOrDefault();
+            DocumentModel document = _context.Documents.Where(d => d.userId == user.Id && d.Id == id).FirstOrDefault();
             return document;
         }
 
@@ -61,13 +56,20 @@ namespace CWDocsCore.Services {
 
 
         public void DeleteDocument(int documentId, string rootPath) {
-            DocumentModel document = _context.Documents.Where(d => d.fileId == documentId).FirstOrDefault();
-            string documentFilePath = Path.Combine(rootPath, _settings["DownloadFilePath"], document.documentName);
+            DocumentModel document = _context.Documents.Where(d => d.Id == documentId).FirstOrDefault();
 
+            string documentFilePath = Path.Combine(rootPath, _settings["DownloadFilePath"], document.documentName);
             var ret = _context.Remove(document);
             _context.SaveChanges();
 
-            System.IO.File.Delete(documentFilePath);
+            if(File.Exists(documentFilePath))
+            {
+                System.IO.File.Delete(documentFilePath);
+            }
+            else
+            {
+                _debugLogger.LogWarning($"WARNING: trying to delete {documentFilePath}, but file doesn't exist.");
+            }
         }
     }
 }
